Transform into a web app using Sinatra (#4)
Transform the existing app into a web application using the Sinatra, Rack, and Puma stack. This will enable building of a true frontend to the backend functionality already in place. As part of this work I added a Rakefile and moved helper scripts into that. I fixed some issues with Rspec configuration not being loaded, which simplified the tests. I also renamed and moved files/directories around to be more idiomatic with respect to the web app stack. No actual routes/controllers are implemented yet, this was just getting the framework in place.