Add schema migration tasks to Rakefile (#10)
Add schema migration and management targets into Rakefile and remove the shell scripts for doing so. The only limitation here is that if you do a complete schema reset then you'll get errors when running the targets because the Sequel ORM requires tables to exist upon loading. Set created/updated timestamps to default to now() to make migrations work from 0 again and also to serve as a backstop for Sequel.