Plant growth and death lifecycle improvements (#12)
Big overhaul of how plants grow, reproduce, and die. Growth and dieback is now based on a parameterized function (EnergyFn) that can be tuned and tested separately. The ability to grow is now focused on the difference between environmental and metabolic energy levels. Plants will die if the size shrinks too much or energy drops to zero. As part of this change I added fields to lifeforms to track the time_stap at which they are created and die. Dead lifeforms are excluded from the standard Environment.lifeform query. I also did a bunch of parameter tuning so that PlantFactory and EnvironmentFactory produce decent growth and reproduction results by default. Lifeforms also has an initial_size column that is used for reproduction.