Improve logging to support multiple log levels (#13)
Migrated to using a simple MyLogger class that supports different log levels and outputs context. This makes it easier to suppress all the trace debugging output that I have added to the code. I opted to roll my own logging class since it was simple enough to write and will give maximum flexibility as the project expands.