Update Lifeforms to use Skills, Params, and Program (#14)
Fundamentally change how lifeforms are modeled to prepare for future improvements to evolution and new species. Different species are no longer sub-classes of Lifeform. Instead, they are modeled as having different Skills, Params, and Programs. E.g. Plants have EnvEnergy (absorb energy from environment), Reproduce, and Grow. The Program consists of Statements and expressions that control how the skills are utilized. Parameters have associated distribution models so we can randomly set and mutate values. This model lays the groundwork for skill acquisition, parameter evolution, and program evolution. All unit tests are passing but there are some bits of functionality that still need to be worked on, including a richer way to write expressions that check on a Lifeform's attributes such as energy and size. But with this commit we have a working app where the Plant Lifeforms endlessly grow and store energy until they get too big and die.