Mitigate unsustainable plant growth by observing environmental energy (#16)
Mitigate unsustainable plant growth by observing environmental energy. Implemented notion of "Observations", which are values that are calculated based on the Lifeform and its Environment. Moved environmental energy  calculations into new Obs::EnvEnergy class so it can be used in Exprs as well as the Skill. Updated Context and Lifeform to support Observations. Update Plant program to grow only if environmental energy is 2x metabolic energy expenditure. This should prevent the plants from growing to an unsustainable size.