Improve logging with more context and terminal colors (#20)
Reimplement logging to support enhanced context and terminal colors. The MyLogger class was refactored into separate conponents (Outputter, Formatter, Message, Router) for greater extensibility. Updated Formatter to recognize when a Lifeform is passed in as context and use this to display identifying details in a consistent manner. This will make it easier to trace application behavior for each lifeform. Added terminal colors to the log output for better visualization. Updated existing application log messages to consistently provide context.