Add file-based logging for the web app and Rakefile (#23)
Update logging setup so that Rakefile and web app both write into the log/ directory in additoin to what they print to stderr. This will make it easier to debug what's going on with Lifeforms in the web app. Rspec tests just print to stderr to avoid cluttering the log with info about temporary objects. As part of this work I updated the "scribe" logging system to include different types of Outputters: Stderr, Stdout, File (with rotation & number of files to keep).