Add Lifeform spawning to the simulation (#25)
Change to dynamic Lifeform spawning rather than rely on up-front specification of population numbers. DB data model is updated with a new table: spawners. A Spawner is parameterized to create Lifeforms for a given Environment and Species. The current behavior supports a minimum and maximum number of Lifeforms of a given species. Beyond that, the Spawner will probabilistically generate new lifeforms each step. 

Other improvements included in this commit:
  * Fix code so that Species are added via migrations and aren't deleted when doing "rake sim:deleteall".
  * Refactor TestFactory.species to take attribute hash like the other factory methods.
  * Attempt to ignore konva in GitHub language stats.