Tune Grazer behavior for more effective Plant consumption (#27)
Tune how a Grazer eats and moves so that it more effectively finds Plants to consume. Update MoveToFood skill so that it will choose the Plant with the most net energy after subtracting off the cost of moving there. This should encourage grazers to move to greener pastures. Note that this does not take into account their metabolic energy, so they may find themselves out of energy while trying to reach it. Update Eat skill so that the grazer selects the Plant within range that has the most energy rather than the closest. This should help resolve the issue where Grazers were dying when surrounded by thriving plants. Adjust Grazer program so that it will only attempt to move if it didn't meet its metabolic energy needs by eating in its current spot. Since Eat chooses the highest energy plant, this will have the effect of keeping the Grazer in place until the nearby plants are eaten down, then it will move to a greener pasture.