Add mutation of a Lifeform's program (#28)
Add the ability for the Lifeform's program to mutate (evolve) as it reproduces. This is a simplistic model that walks the tree of Exprs and probabilistically applies changes to them, e.g. swapping out a skill, updating a numeric constant, changing if() to seq(), and so on. The changes can lead to non-sensical programs to be sure, and likely some tuning is needed to help evolution proceed in promising directions. But this at least lays the foundation for how this can work. The testing relies on a lot of random trials to ensure that the mutated expressions still return expected values and don't raise exceptions.